requirements.txt
alembic==0.9.5
bcrypt==3.1.3
cffi==1.10.0
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.0
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3
pycparser==2.18
PyJWT==1.5.2
python-dateutil==2.6.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.13
validate-email==1.3
Werkzeug==0.12.2
